feat: includes artists in collection during inquiry
