+++ src/Components/Inquiry/Hooks/useInquiryContext.tsx
@@ -16,21 +16,16 @@ import { SystemQueryRenderer } from "System/Relay/SystemQueryRenderer"
 import { useSystemContext } from "System/Hooks/useSystemContext"
 import { Visited } from "Components/Inquiry/Visited"
 import { logger } from "Components/Inquiry/util"
-import { Location } from "Components/LocationAutocompleteInput"
 import { Spinner } from "@artsy/palette"
 import { Environment as IEnvironment } from "react-relay"
 import { setAfterAuthAction } from "Utils/Hooks/useAuthIntent"
 
-export type Context = {
+export interface Context
+  extends Omit<useInquiryContext_me$data, " $fragmentType"> {
   askSpecialist: boolean
-  collectorLevel?: number | null
   enableCreateAlert: boolean
   isLoggedIn: boolean
-  location?: Location | null
-  otherRelevantPositions?: string | null
-  profession?: string | null
   requiresReload: boolean
-  shareFollows: boolean
 }
 
 export const DEFAULT_CONTEXT: Context = {
@@ -43,6 +38,8 @@ export const DEFAULT_CONTEXT: Context = {
   profession: null,
   requiresReload: false,
   shareFollows: false,
+  userInterestsConnection: { totalCount: 0 },
+  collectorProfile: { lastUpdatePromptAt: null },
 }
 
 export const DEFAULT_MESSAGE =
@@ -210,14 +207,7 @@ const InquiryContextContext: React.FC<InquiryContextContextProps> = ({
   const { setContext } = useInquiryContext()
 
   useEffect(() => {
-    setContext({
-      collectorLevel: me?.collectorLevel,
-      isLoggedIn: !!me,
-      location: !!me?.location?.city ? { city: me.location.city } : null,
-      otherRelevantPositions: me?.otherRelevantPositions,
-      profession: me?.profession,
-      shareFollows: !!me?.shareFollows,
-    })
+    setContext({ ...me, isLoggedIn: !!me })
   }, [me, setContext])
 
   return <>{children}</>
@@ -231,10 +221,19 @@ const InquiryContextContextFragmentContainer = createFragmentContainer(
         collectorLevel
         location {
           city
+          state
+          postalCode
+          country
         }
         otherRelevantPositions
         profession
         shareFollows
+        userInterestsConnection(interestType: ARTIST, first: 1) {
+          totalCount
+        }
+        collectorProfile {
+          lastUpdatePromptAt
+        }
       }
     `,
   }
