+++ src/Components/Inquiry/Views/InquiryLogin.tsx
@@ -82,7 +82,6 @@ export const InquiryLogin: React.FC = () => {
         location: {
           city: user.location?.city,
           state: user.location?.state,
-          stateCode: user.location?.stateCode,
           postalCode: user.location?.postalCode,
           country: user.location?.country,
         },
